improve file sctructure
